moved settings enums to settings header
