+++ AP_Randomizer/include/Engine.hpp
@@ -2,7 +2,7 @@
 #include <variant>
 #include "Unreal/UObject.hpp"
 #include "GameData.hpp"
-#include "APOptions.hpp"
+#include "Settings.hpp"
 
 namespace Engine {
 	using RC::Unreal::UObject;
@@ -33,7 +33,7 @@ namespace Engine {
 	void PrintToConsole(std::wstring, std::wstring);
 	void PrintToConsole(std::wstring);
 	void ShowPopup(std::variant<std::wstring, ItemPopup>);
-	void UpdatePopupDisplay(EPopupDisplay::Type);
+	void UpdatePopupDisplay(Settings::EPopupDisplay::Type);
 	void HealPlayer();
 	void GivePlayerPower();
 	void WarpToSpawn();
